Move 1.x API into subdirectory
